+++ hydra-node/src/Hydra/Network.hs
@@ -55,7 +55,9 @@ instance Contravariant (Network m) where
 type NetworkCallback msg m = msg -> m ()
 
 -- | A type tying both inbound and outbound messages sending in a single /Component/.
-type NetworkComponent m msg a = NetworkCallback msg m -> (Network m msg -> m a) -> m a
+--
+-- A `NetworkComponent` can have different inbound and outbound message types.
+type NetworkComponent m inbound outbound a = NetworkCallback inbound m -> (Network m outbound -> m a) -> m a
 
 -- * Types used by concrete implementations