+++ hydra-node/src/Hydra/Network/Ouroboros.hs
@@ -7,6 +7,7 @@ module Hydra.Network.Ouroboros (
   withOuroborosNetwork,
   withIOManager,
   TraceOuroborosNetwork,
+  WithHost,
   module Hydra.Network,
   encodeTraceSendRecvFireForget,
 ) where
@@ -124,9 +125,7 @@ withOuroborosNetwork ::
   Tracer IO (WithHost (TraceOuroborosNetwork msg)) ->
   Host ->
   [Host] ->
-  NetworkCallback msg IO ->
-  (Network IO msg -> IO ()) ->
-  IO ()
+  NetworkComponent IO msg msg ()
 withOuroborosNetwork tracer localHost remoteHosts networkCallback between = do
   bchan <- newBroadcastTChanIO
   let newBroadcastChannel = atomically $ dupTChan bchan